requirements.txt
altair
attrs
avro-python3
blinker
boto3
botocore
cachetools
certifi
charset-normalizer
click
confluent-kafka
docopt
Faker
gitdb
GitPython
idna
Jinja2
jmespath
jsonschema
jsonschema-specifications
kafka-python
MarkupSafe
names
narwhals
numpy
packaging
pandas
pillow
pipreqs
plotly
protobuf
py4j
pyarrow
pydeck
pyspark==3.4.0
python-dateutil
python-dotenv
pytz
referencing
requests
rpds-py
s3transfer
six
smmap
streamlit
tenacity
toml
tornado
tzdata
urllib3
yarg
